# openapi.yaml
openapi: 3.1.0
...
info:
  title: KATHER 
  version: 1.0.0
  description: >-
    KATHER - Arabic Poetry API - serves curated Arabic poetry data, including poets, poems,
    verses, eras, topics, meters, and rhyme schemes. It is designed for developers
    building Arabic literature apps, NLP tools, educational platforms, and cultural
    archives. All Arabic text is UTF-8 encoded and should be displayed right-to-left.
  termsOfService: https://example.com/terms
  contact:
    email: support@example.com
    url: https://example.com/support
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
  x-logo:
    url: https://example.com/assets/arabic-poetry-logo.png
    backgroundColor: "#ffffff"
    altText: KATHER
servers:
  - url: http://localhost:{port}{basePath}
    description: Local development
    variables:
      port:
        default: "3000"
      basePath:
        default: "/v1"
        enum:
          - ""
          - "/v1"
  - url: https://staging.api.example.com{basePath}
    description: Staging
    variables:
      basePath:
        default: "/v1"
        enum:
          - ""
          - "/v1"
  - url: https://api.example.com{basePath}
    description: Production
    variables:
      basePath:
        default: "/v1"
        enum:
          - ""
          - "/v1"
security: []
...
paths:
...
  /catalog:
    get:
      operationId: getCatalog
      summary: Get catalog overview
      description: >-
        Returns all catalog collections (eras, countries, topics, seas, quawafi)
        used for filtering poets, poems, and lines. Arabic labels are UTF-8 encoded
        and should be rendered RTL.
      tags: [Catalog]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      responses:
        "200":
          description: Catalog collection payload.
          headers:
            Cache-Control:
              description: Cache policy for catalog data.
              schema:
                type: string
                example: public, max-age=86400, stale-while-revalidate=604800
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogCollectionResponse"
              examples:
                catalog:
                  value:
                    status: success
                    data:
                      eras:
                        - id: abbasid-era
                          labelAr: "العصر العباسي"
                          aliases: []
                      countries:
                        - id: iraq
                          labelAr: "العراق"
                          aliases: []
                      quawafi:
                        - id: q-ل
                          labelAr: "قافية اللام (ل)"
                          aliases: []
                      seas:
                        - id: el-tawil
                          labelAr: "الطويل"
                          aliases: []
                      topics:
                        - id: pride
                          labelAr: "قصائد فخر"
                          aliases: []
                    meta:
                      collections: 5
                      counts:
                        eras: 11
                        countries: 21
                        quawafi: 33
                        seas: 40
                        topics: 28
                    links:
                      self: /v1/catalog
                      eras: /v1/catalog/eras
                      countries: /v1/catalog/countries
                      quawafi: /v1/catalog/quawafi
                      seas: /v1/catalog/seas
                      topics: /v1/catalog/topics
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/catalog"
        - lang: JavaScript
          label: fetch
          source: |-
            const res = await fetch("https://api.example.com/v1/catalog");
            console.log(await res.json());
        - lang: Python
          label: requests
          source: |-
            import requests
            print(requests.get("https://api.example.com/v1/catalog").json())
...
  /poets:
    get:
      operationId: listPoets
      summary: List poets
      description: >-
        Returns a paginated list of poets with optional filtering by era, country,
        gender, and search query. Arabic names are UTF-8 encoded and should be
        rendered RTL. Rate limiting applies to this endpoint.
      tags: [Poets]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      parameters:
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitPoetsParam"
        - $ref: "#/components/parameters/MetaParam"
        - $ref: "#/components/parameters/EraParam"
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/GenderParam"
        - $ref: "#/components/parameters/SearchQueryShortParam"
        - $ref: "#/components/parameters/PoetSortParam"
      responses:
        "200":
          description: Paginated list of poets.
          headers:
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoetListResponse"
              examples:
                poets:
                  value:
                    status: success
                    data:
                      - id: 1
                        engName: Al-Mutanabbi
                        arabName: "المتنبي"
                        engEra: abbasid-era
                        arabEra: "العصر العباسي"
                        engCountry: iraq
                        arabCountry: "العراق"
                        gender: male
                        created_at: "2026-01-01T00:00:00.000Z"
                        poem_count: 128
                    pagination:
                      offset: 0
                      page: 1
                      limit: 50
                      total: 345
                      total_pages: 7
                      has_more: true
                    links:
                      self: /v1/poets?limit=50&offset=0
                      next: /v1/poets?limit=50&offset=50
                      prev: null
                      random: /v1/poets/random
                    meta:
                      filters:
                        era: abbasid-era
                        country: iraq
                        gender: male
                        q: "متنبي"
                        sort: name
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/poets?limit=5&offset=0&era=abbasid-era"
        - lang: JavaScript
          label: fetch
          source: |-
            const url = new URL("https://api.example.com/v1/poets");
            url.searchParams.set("limit", "5");
            url.searchParams.set("offset", "0");
            const res = await fetch(url);
            console.log(await res.json());
        - lang: Python
          label: requests
          source: |-
            import requests
            resp = requests.get("https://api.example.com/v1/poets", params={"limit": 5, "offset": 0})
            print(resp.json())
...
  /poets/{id}:
    get:
      operationId: getPoetById
      summary: Get poet by id
      description: >-
        Returns a poet profile by numeric id. Arabic fields are UTF-8 encoded and
        should be rendered RTL.
      tags: [Poets]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Poet profile.
          headers:
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoetDetailResponse"
              examples:
                poet:
                  value:
                    status: success
                    data:
                      id: 1
                      engName: Al-Mutanabbi
                      arabName: "المتنبي"
                      bio: "أبو الطيب المتنبي من أبرز شعراء العربية في العصر العباسي."
                      engEra: abbasid-era
                      arabEra: "العصر العباسي"
                      engCountry: iraq
                      arabCountry: "العراق"
                      gender: male
                      created_at: "2026-01-01T00:00:00.000Z"
                      poem_count: 128
                    links:
                      self: /v1/poets/1
                      poems: /v1/poets/1/poems
                      stats: /v1/poets/1/stats
                      random: /v1/poets/random
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/poets/1"
        - lang: JavaScript
          label: fetch
          source: |-
            const res = await fetch("https://api.example.com/v1/poets/1");
            console.log(await res.json());
        - lang: Python
          label: requests
          source: |-
            import requests
            print(requests.get("https://api.example.com/v1/poets/1").json())
...
  /poems:
    get:
      operationId: listPoems
      summary: List poems
      description: >-
        Returns a paginated list of poems with optional filtering by poet, era,
        country, gender, quafia, sea, topic, and search query. Arabic fields are
        UTF-8 encoded and should be rendered RTL.
      tags: [Poems]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      parameters:
        - $ref: "#/components/parameters/LimitPoemsParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/MetaParam"
        - $ref: "#/components/parameters/PoetIdQueryParam"
        - $ref: "#/components/parameters/EraParam"
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/GenderParam"
        - $ref: "#/components/parameters/QuafiaParam"
        - $ref: "#/components/parameters/SeaParam"
        - $ref: "#/components/parameters/TopicParam"
        - $ref: "#/components/parameters/PoemTypeParam"
        - $ref: "#/components/parameters/SearchQueryShortParam"
      responses:
        "200":
          description: Paginated list of poems.
          headers:
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoemListResponse"
              examples:
                poems:
                  value:
                    status: success
                    data:
                      - id: 101
                        name: "قصيدة الخيل والليل"
                        engTopic: Pride
                        arabTopic: "قصائد فخر"
                        type: classical
                        engSea: el-tawil
                        arabSea: "الطويل"
                        quafia: q-ل
                        order: 1
                        created_at: "2026-01-01T00:00:00.000Z"
                        poetId: 1
                        line_count: 45
                        poet:
                          id: 1
                          engName: Al-Mutanabbi
                          arabName: "المتنبي"
                          engEra: abbasid-era
                          arabEra: "العصر العباسي"
                          engCountry: iraq
                          arabCountry: "العراق"
                          gender: male
                          created_at: "2026-01-01T00:00:00.000Z"
                          poem_count: 128
                    pagination:
                      offset: 0
                      page: 1
                      limit: 20
                      total: 540
                      total_pages: 27
                      has_more: true
                    links:
                      self: /v1/poems?limit=20&offset=0
                      next: /v1/poems?limit=20&offset=20
                      prev: null
                      random: /v1/poems/random
                    meta:
                      filters:
                        poetId: 1
                        era: abbasid-era
                        country: iraq
                        gender: male
                        quafia: q-ل
                        sea: el-tawil
                        topic: Pride
                        type: classical
                        q: "ليل"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/poems?limit=5&offset=0&poetId=1"
        - lang: JavaScript
          label: fetch
          source: |-
            const url = new URL("https://api.example.com/v1/poems");
            url.searchParams.set("poetId", "1");
            const res = await fetch(url);
            console.log(await res.json());
        - lang: Python
          label: requests
          source: |-
            import requests
            print(requests.get("https://api.example.com/v1/poems", params={"poetId": 1}).json())
...
  /poems/{id}:
    get:
      operationId: getPoemById
      summary: Get poem by id
      description: >-
        Returns a poem and its lines with pagination. Arabic fields and line content
        are UTF-8 encoded and should be rendered RTL.
      tags: [Poems]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - $ref: "#/components/parameters/LimitPoemDetailLinesParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: Poem detail with lines.
          headers:
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoemDetailResponse"
              examples:
                poemDetail:
                  value:
                    status: success
                    data:
                      poem:
                        id: 101
                        name: "قصيدة الخيل والليل"
                        engTopic: Pride
                        arabTopic: "قصائد فخر"
                        type: classical
                        engSea: el-tawil
                        arabSea: "الطويل"
                        quafia: q-ل
                        order: 1
                        created_at: "2026-01-01T00:00:00.000Z"
                        poetId: 1
                        line_count: 45
                        poet:
                          id: 1
                          engName: Al-Mutanabbi
                          arabName: "المتنبي"
                          engEra: abbasid-era
                          arabEra: "العصر العباسي"
                          engCountry: iraq
                          arabCountry: "العراق"
                          gender: male
                          created_at: "2026-01-01T00:00:00.000Z"
                          poem_count: 128
                      lines:
                        - id: 1001
                          content: "الخيل والليل والبيداء تعرفني"
                          type: 1
                          order: 1
                        - id: 1002
                          content: "والسيف والرمح والقرطاس والقلم"
                          type: 1
                          order: 2
                      pagination:
                        offset: 0
                        page: 1
                        limit: 12
                        total: 45
                        total_pages: 4
                        has_more: true
                    links:
                      self: /v1/poems/101?limit=12&offset=0
                      lines: /v1/poems/101/lines?limit=12&offset=0
                      context: /v1/poems/101/context
                      random: /v1/poems/random
                      poet: /v1/poets/1
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/poems/101?limit=12&offset=0"
        - lang: JavaScript
          label: fetch
          source: |-
            const res = await fetch("https://api.example.com/v1/poems/101?limit=12&offset=0");
            console.log(await res.json());
        - lang: Python
          label: requests
          source: |-
            import requests
            print(requests.get("https://api.example.com/v1/poems/101", params={"limit": 12, "offset": 0}).json())
...
  /lines:
    get:
      operationId: listLines
      summary: List lines
      description: >-
        Returns a paginated list of poem lines with optional filters for poet, era,
        country, gender, quafia, sea, topic, and line type. Arabic line content is
        UTF-8 encoded and should be rendered RTL.
      tags: [Lines]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      parameters:
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitLinesParam"
        - $ref: "#/components/parameters/MetaParam"
        - $ref: "#/components/parameters/PoetIdQueryParam"
        - $ref: "#/components/parameters/EraParam"
        - $ref: "#/components/parameters/CountryParam"
        - $ref: "#/components/parameters/GenderParam"
        - $ref: "#/components/parameters/QuafiaParam"
        - $ref: "#/components/parameters/SeaParam"
        - $ref: "#/components/parameters/TopicParam"
        - $ref: "#/components/parameters/LineTypeParam"
        - $ref: "#/components/parameters/SearchQueryLongParam"
      responses:
        "200":
          description: Paginated list of lines.
          headers:
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineListResponse"
              examples:
                lines:
                  value:
                    status: success
                    data:
                      - id: 1001
                        content: "الخيل والليل والبيداء تعرفني"
                        type: 1
                        order: 1
                        created_at: "2026-01-01T00:00:00.000Z"
                        poemId: 101
                        poem:
                          id: 101
                          name: "قصيدة الخيل والليل"
                          engTopic: Pride
                          arabTopic: "قصائد فخر"
                          type: classical
                          engSea: el-tawil
                          arabSea: "الطويل"
                          quafia: q-ل
                          order: 1
                          created_at: "2026-01-01T00:00:00.000Z"
                          poetId: 1
                          line_count: 45
                          poet:
                            id: 1
                            engName: Al-Mutanabbi
                            arabName: "المتنبي"
                            engEra: abbasid-era
                            arabEra: "العصر العباسي"
                            engCountry: iraq
                            arabCountry: "العراق"
                            gender: male
                            created_at: "2026-01-01T00:00:00.000Z"
                            poem_count: 128
                    pagination:
                      offset: 0
                      page: 1
                      limit: 50
                      total: 4600
                      total_pages: 92
                      has_more: true
                    links:
                      self: /v1/lines?limit=50&offset=0
                      next: /v1/lines?limit=50&offset=50
                      prev: null
                      random: /v1/lines/random
                    meta:
                      filters:
                        poemId: null
                        poetId: 1
                        era: abbasid-era
                        country: iraq
                        gender: male
                        quafia: q-ل
                        sea: el-tawil
                        topic: Pride
                        lineType: 1
                        q: "الخيل"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/lines?limit=5&offset=0&poetId=1"
        - lang: JavaScript
          label: fetch
          source: |-
            const res = await fetch("https://api.example.com/v1/lines?limit=5&offset=0&poetId=1");
            console.log(await res.json());
        - lang: Python
          label: requests
          source: |-
            import requests
            print(requests.get("https://api.example.com/v1/lines", params={"poetId": 1}).json())
...
  /lines/{id}:
    get:
      operationId: getLineById
      summary: Get line by id
      description: >-
        Returns a single line by numeric id. Arabic content is UTF-8 encoded and
        should be rendered RTL.
      tags: [Lines]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Line detail.
          headers:
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineDetailResponse"
              examples:
                line:
                  value:
                    status: success
                    data:
                      id: 1001
                      content: "الخيل والليل والبيداء تعرفني"
                      type: 1
                      order: 1
                      created_at: "2026-01-01T00:00:00.000Z"
                      poemId: 101
                      poem:
                        id: 101
                        name: "قصيدة الخيل والليل"
                        engTopic: Pride
                        arabTopic: "قصائد فخر"
                        type: classical
                        engSea: el-tawil
                        arabSea: "الطويل"
                        quafia: q-ل
                        order: 1
                        created_at: "2026-01-01T00:00:00.000Z"
                        poetId: 1
                        line_count: 45
                        poet:
                          id: 1
                          engName: Al-Mutanabbi
                          arabName: "المتنبي"
                          engEra: abbasid-era
                          arabEra: "العصر العباسي"
                          engCountry: iraq
                          arabCountry: "العراق"
                          gender: male
                          created_at: "2026-01-01T00:00:00.000Z"
                          poem_count: 128
                    links:
                      self: /v1/lines/1001
                      poem: /v1/poems/101
                      poet: /v1/poets/1
                      random: /v1/lines/random
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/lines/1001"
        - lang: JavaScript
          label: fetch
          source: |-
            const res = await fetch("https://api.example.com/v1/lines/1001");
            console.log(await res.json());
        - lang: Python
          label: requests
          source: |-
            import requests
            print(requests.get("https://api.example.com/v1/lines/1001").json())
  /openapi.yaml:
    get:
      operationId: getOpenApiYaml
      summary: Get OpenAPI YAML
      description: Returns the raw OpenAPI YAML specification.
      tags: [Docs]
      x-rateLimit:
        limit: 100
        windowMs: 900000
        policy: fixed-window
      responses:
        "200":
          description: OpenAPI YAML document.
          headers:
            Content-Type:
              description: YAML content type.
              schema:
                type: string
                example: application/yaml
            X-API-Version:
              $ref: "#/components/headers/X-API-Version"
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/yaml:
              schema:
                type: string
                description: OpenAPI YAML content.
                example: "openapi: 3.1.0"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalServerError"
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |-
            curl -sS "https://api.example.com/v1/openapi.yaml"
        - lang: JavaScript
          label: fetch
          source: |-
            const res = await fetch("https://api.example.com/v1/openapi.yaml");
            console.log(await res.text());
        - lang: Python
          label: requests
          source: |-
            import requests
            print(requests.get("https://api.example.com/v1/openapi.yaml").text)
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Basic auth for /metrics when enabled by env vars.
  headers:
    X-API-Version:
      description: Semantic API version.
      schema:
        type: string
        example: 1.0.0
    X-Request-Id:
      description: Unique request correlation ID.
      schema:
        type: string
        example: "01HZX8JQG5T4XK3BPCQ1V6Y9AE"
...
  schemas:
...
    CatalogCollectionResponse:
      type: object
      description: Catalog overview response.
      properties:
        status:
          type: string
          description: Response status.
          example: success
        data:
          $ref: "#/components/schemas/CatalogCollection"
        meta:
          type: object
          description: Collection metadata.
          properties:
            collections:
              type: integer
              description: Number of catalog collections.
              example: 5
            counts:
              $ref: "#/components/schemas/CatalogCounts"
          required: [collections, counts]
        links:
          type: object
          description: Collection links.
          properties:
            self:
              type: string
              description: Catalog root.
              example: /v1/catalog
            eras:
              type: string
              description: Eras collection.
              example: /v1/catalog/eras
            countries:
              type: string
              description: Countries collection.
              example: /v1/catalog/countries
            quawafi:
              type: string
              description: Quawafi collection.
              example: /v1/catalog/quawafi
            seas:
              type: string
              description: Seas collection.
              example: /v1/catalog/seas
            topics:
              type: string
              description: Topics collection.
              example: /v1/catalog/topics
          required: [self, eras, countries, quawafi, seas, topics]
      required: [status, data, meta, links]
...